Implement Comparative Analysis of Spectral vs. Spatial Autoencoders with Training
This commit introduces two new Autoencoder architectures, a benchmarking suite, and a training example to compare the "Zenith" Spectral approach against a standard "Spatial" Convolutional baseline.

1.  **New Models (`include/dreidel/models/ComparativeAE.hpp`):**
    *   `ZenithHierarchicalAE`: Hierarchical spectral autoencoder utilizing `PixelUnshuffle` and `ZenithBlock`.
    *   `ConvBaselineAE`: Convolutional baseline using strided `Conv2D` and `ResidualBlock`.

2.  **New Layers:**
    *   `PixelShuffle.hpp`: Implements `PixelShuffle` and `PixelUnshuffle` (NHWC layout).
    *   `ConvTranspose2D.hpp`: Optimized implementation for `kernel_size == stride` using `Conv2D(1x1)` + `PixelShuffle`.

3.  **Utilities & Optimizers:**
    *   `include/dreidel/utils/WaveletGen2D.hpp`: Generates diverse 2D wavelet patterns (Gabor, Mexican Hat, etc.).
    *   `include/dreidel/optim/SimpleAdam.hpp`: A lightweight Adam optimizer implementation.

4.  **Fixes:**
    *   Updated `Conv2D.hpp` to use `long` for spatial coordinate calculations, preventing `size_t` underflow when using padding.

5.  **Examples:**
    *   `examples/train_comparative_ae.cpp`: A training loop comparing both models on 32x32 wavelet images.
    *   `benchmarks/benchmark_comparative_ae.cpp`: Performance benchmarking script.

+++ include/dreidel/layers/Conv2D.hpp
@@ -63,18 +63,19 @@ public:
             for(size_t h_out=0; h_out<H_out; ++h_out) {
                 for(size_t w_out=0; w_out<W_out; ++w_out) {
 
-                    size_t h_in_start = h_out * stride_ - padding_;
-                    size_t w_in_start = w_out * stride_ - padding_;
+                    // Use long to prevent underflow during subtraction of padding
+                    long h_in_start = static_cast<long>(h_out * stride_) - static_cast<long>(padding_);
+                    long w_in_start = static_cast<long>(w_out * stride_) - static_cast<long>(padding_);
 
                     for(size_t c_out=0; c_out<C_out; ++c_out) {
                         T acc = b_ptr[c_out];
 
                         for(size_t ky=0; ky<kernel_size_; ++ky) {
                             for(size_t kx=0; kx<kernel_size_; ++kx) {
-                                int h_in = h_in_start + ky;
-                                int w_in = w_in_start + kx;
+                                long h_in = h_in_start + ky;
+                                long w_in = w_in_start + kx;
 
-                                if (h_in >= 0 && h_in < H && w_in >= 0 && w_in < W) {
+                                if (h_in >= 0 && h_in < static_cast<long>(H) && w_in >= 0 && w_in < static_cast<long>(W)) {
                                     for(size_t c_in=0; c_in<C_in; ++c_in) {
                                         // w index: c_out * (C_in*K*K) + c_in * (K*K) + ky*K + kx
                                         // in index: ((n*H + h_in)*W + w_in)*C_in + c_in
@@ -142,18 +143,18 @@ public:
             for(size_t h_out=0; h_out<H_out; ++h_out) {
                 for(size_t w_out=0; w_out<W_out; ++w_out) {
 
-                    int h_in_start = h_out * stride_ - padding_;
-                    int w_in_start = w_out * stride_ - padding_;
+                    long h_in_start = static_cast<long>(h_out * stride_) - static_cast<long>(padding_);
+                    long w_in_start = static_cast<long>(w_out * stride_) - static_cast<long>(padding_);
 
                     for(size_t c_out=0; c_out<out_channels_; ++c_out) {
                         T dy = go_ptr[((n*H_out+h_out)*W_out+w_out)*out_channels_ + c_out];
 
                         for(size_t ky=0; ky<kernel_size_; ++ky) {
                             for(size_t kx=0; kx<kernel_size_; ++kx) {
-                                int h_in = h_in_start + ky;
-                                int w_in = w_in_start + kx;
+                                long h_in = h_in_start + ky;
+                                long w_in = w_in_start + kx;
 
-                                if (h_in >= 0 && h_in < H && w_in >= 0 && w_in < W) {
+                                if (h_in >= 0 && h_in < static_cast<long>(H) && w_in >= 0 && w_in < static_cast<long>(W)) {
                                     for(size_t c_in=0; c_in<C_in; ++c_in) {
                                         size_t w_idx = ((c_out * in_channels_ + c_in) * kernel_size_ + ky) * kernel_size_ + kx;
                                         size_t in_idx = ((n*H + h_in)*W + w_in)*C_in + c_in;
